Ajuste nas coordenadas que devem aceitar numeros com ponto flutuante
git-svn-id: http://svn.infox.com.br/repos/infoxnovoepa/aplicacao/trunk/epa2@544 069f99d3-da65-4ab0-85f1-cf03781be9e7

+++ src/main/java/br/com/infox/ibpm/xpdl/element/GraphicsInfoXPDL.java
@@ -19,12 +19,12 @@ public class GraphicsInfoXPDL implements Serializable {
 	
 	public static GraphicsInfoXPDL createInstance(Element element) {
 		Element info = XmlUtil.getChildByIndex(element, "NodeGraphicsInfo", 0);
-		double height = Integer.parseInt(info.getAttributeValue("Height"));
-		double width = Integer.parseInt(info.getAttributeValue("Width"));
+		double height = Double.parseDouble(info.getAttributeValue("Height"));
+		double width = Double.parseDouble(info.getAttributeValue("Width"));
 		
 		Element coordinates = XmlUtil.getChildByIndex(info, "Coordinates", 0);
-		double x = Float.parseFloat(coordinates.getAttributeValue("XCoordinate"));
-		double y = Float.parseFloat(coordinates.getAttributeValue("YCoordinate"));
+		double x = Double.parseDouble(coordinates.getAttributeValue("XCoordinate"));
+		double y = Double.parseDouble(coordinates.getAttributeValue("YCoordinate"));
 		Point2D position = new Point2D.Double(x,y);
 		
 		return new GraphicsInfoXPDL(position, width, height);
